requirements.txt
streamlit
pdfplumber
python-dotenv
langchain
langchain-groq
tavily-python
